refactor: moved helper functions from customFetchBase.ts to utils.ts
