+++ src/api/customFetchBase.ts
@@ -1,4 +1,3 @@
-import { SerializedError } from '@reduxjs/toolkit'
 import {
   BaseQueryFn,
   FetchArgs,
@@ -12,42 +11,10 @@ import { API_URL } from '../constants'
 import { httpOnlyProxy } from '../env'
 import { updateCurrentUser } from '../slices/currentUserSlice'
 import { RootState } from '../store'
-import { RefreshTokenResponse } from '../types'
-import { authApi } from './auth.api'
+import { AddTokenToUrl, runRefreshToken, updateTokenInArgs } from './utils'
     
 // Create a new mutex
 const mutex = new Mutex()
-
-const getQueryPath = (url: string) => {
-  const matchResult = url.match(/(^.*auth=)(.*)$/)
-  return matchResult ? matchResult[1] : ''
-} 
-
-const updateTokenInArgs = (args: string | FetchArgs, newToken: string) => {
-  if (typeof args === 'string') {
-    return getQueryPath(args) + newToken
-  } else {
-    args.url = getQueryPath(args.url) + newToken
-    return args
-  }
-}
-
-const AddTokenToUrl = (args: string | FetchArgs, token: string) => {
-  const queryString = '?auth=' + token
-  if (typeof args === 'string') {
-    return args + queryString
-  } else {
-    args.url = args.url + queryString
-    return args
-  }
-}
-
-const runRefreshToken = async (api: any, refreshToken: string) => {
-  const res: {
-    data: RefreshTokenResponse } | { error: FetchBaseQueryError | SerializedError
-    } = await api.dispatch(authApi.endpoints.refreshToken.initiate(refreshToken))
-  return res
-}
 
 const baseQuery = fetchBaseQuery({
   baseUrl: API_URL + '/users'
@@ -73,10 +40,6 @@ const customFetchBase: BaseQueryFn<
 
     if (!mutex.isLocked()) {
       const release = await mutex.acquire()
-
-      // let res: {
-      //     data: RefreshTokenResponse } | { error: FetchBaseQueryError | SerializedError
-      //   } = undefined
 
       let submitToken: string = ''
